Revert "Update Github_Tutorial.ino"
This reverts commit e7c25a0038005218b5ec1b9c94cfda6c05198591.

+++ Github_Tutorial.ino
@@ -21,7 +21,7 @@ void setup()
 
 void loop() 
 {
-  int myValue = 0;
+  byte myValue = 0;
   myValue = analogRead(A0);
   
   Serial.print("The value is: ");
@@ -29,3 +29,4 @@ void loop()
 
   delay(250);
 }
+
